Added documentation for handler building logic
